Refactor EmptyEGraphTest into base class and implementation
